refactor(field.unit-test.ts): remove rotation test from field.unit-test.ts
